feat:add-regras-de-negocio

+++ src/main/java/br/com/jovemtech/productordermanager/usecase/pedido/RemoverPedidoPorIdUC.java
@@ -2,6 +2,9 @@ package br.com.jovemtech.productordermanager.usecase.pedido;
 
 import br.com.jovemtech.productordermanager.config.exception.ResourceNotFoundException;
 import br.com.jovemtech.productordermanager.infrastructure.repository.PedidoRepository;
+import br.com.jovemtech.productordermanager.infrastructure.repository.ProdutoRepository;
+import br.com.jovemtech.productordermanager.schema.PedidoSchema;
+import br.com.jovemtech.productordermanager.schema.StatusPedido;
 import lombok.RequiredArgsConstructor;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
@@ -13,14 +16,29 @@ import org.springframework.transaction.annotation.Transactional;
 public class RemoverPedidoPorIdUC {
 
     private final PedidoRepository pedidoRepository;
+    private final ProdutoRepository produtoRepository;
 
-    @Transactional(propagation = Propagation.SUPPORTS)
-    public void remover(Long id) {
+    @Transactional(propagation = Propagation.REQUIRED)
+    public void execute(Long id) {
         if(!pedidoRepository.existsById(id)) {
-            throw new ResourceNotFoundException("Erro ao buscar produto com o id " + id);
+            throw new ResourceNotFoundException("Erro ao buscar pedido com o id " + id);
         }
         try{
-            pedidoRepository.deleteById(id);
+            PedidoSchema pedido = pedidoRepository.getReferenceById(id);
+
+            if(pedido.getStatus() != StatusPedido.PAGO
+                    && pedido.getStatus() != StatusPedido.FINALIZADO){
+
+                pedido.getItens().forEach(itemPedido -> {
+                    itemPedido.getId().getProduto().setQuantidadeEstoque(
+                            itemPedido.getId().getProduto().getQuantidadeEstoque() + itemPedido.getQuantidade()
+                    );
+                    produtoRepository.save(itemPedido.getId().getProduto());
+                });
+
+                pedidoRepository.deleteById(id);
+
+            }
         } catch (DataIntegrityViolationException e) {
             throw new RuntimeException(e.getMessage());
         }
